Add files via upload
This blood bank is built using PHP, JavaScript, Bootstrap, and CSS. Talking about the features of this system, it contains both the hospital user and the blood receiver user. All aspects of managing donors, blood needs, and editing places are handled by the hospital section. The receivers have the option to request the necessary blood sample.

In this instance, the skills utilized include HTML, CSS, Bootstrap, JavaScript as a frontend, PHP as a backend, and MySQL for the database.